Make general package struct? Probably no, because we don't want to accidentally mix these. I'm going to not do what's in this commit
