+++ src/npm.rs
@@ -430,173 +430,6 @@ fn part(input: &str) -> ParseResult<&str> {
   )(input)
 }
 
-#[derive(Debug, Error, Clone)]
-#[error("Invalid npm package name and version reference '{text}'. {message}")]
-pub struct NpmPackageNvReferenceParseError {
-  pub message: String,
-  pub text: String,
-}
-
-/// A npm package name and version with a potential subpath.
-#[derive(
-  Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash, Serialize, Deserialize,
-)]
-pub struct NpmPackageNvReference {
-  pub nv: NpmPackageNv,
-  pub sub_path: Option<String>,
-}
-
-impl NpmPackageNvReference {
-  pub fn from_specifier(
-    specifier: &Url,
-  ) -> Result<Self, NpmPackageNvReferenceParseError> {
-    Self::from_str(specifier.as_str())
-  }
-
-  #[allow(clippy::should_implement_trait)]
-  pub fn from_str(nv: &str) -> Result<Self, NpmPackageNvReferenceParseError> {
-    use monch::*;
-
-    fn sub_path(input: &str) -> ParseResult<&str> {
-      let (input, _) = ch('/')(input)?;
-      Ok(("", input))
-    }
-
-    fn parse_ref(input: &str) -> ParseResult<NpmPackageNvReference> {
-      let (input, _) = tag("npm:")(input)?;
-      let (input, nv) = parse_nv(input)?;
-      let (input, maybe_sub_path) = maybe(sub_path)(input)?;
-      Ok((
-        input,
-        NpmPackageNvReference {
-          nv,
-          sub_path: maybe_sub_path.map(ToOwned::to_owned),
-        },
-      ))
-    }
-
-    with_failure_handling(parse_ref)(nv).map_err(|err| {
-      NpmPackageNvReferenceParseError {
-        message: format!("{err:#}"),
-        text: nv.to_string(),
-      }
-    })
-  }
-
-  pub fn as_specifier(&self) -> Url {
-    let version_text = self.nv.version.to_string();
-    let capacity = 5 /* npm:/ */
-      + self.nv.name.len()
-      + 1 /* @ */
-      + version_text.len()
-      + self.sub_path.as_ref().map(|p| p.len() + 1 /* slash  */).unwrap_or(0);
-    let mut text = String::with_capacity(capacity);
-    text.push_str("npm:/");
-    text.push_str(&self.nv.name);
-    text.push('@');
-    text.push_str(&version_text);
-    if let Some(sub_path) = &self.sub_path {
-      text.push('/');
-      text.push_str(sub_path);
-    }
-    debug_assert_eq!(text.len(), capacity);
-    Url::parse(&text).unwrap()
-  }
-}
-
-impl std::fmt::Display for NpmPackageNvReference {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    if let Some(sub_path) = &self.sub_path {
-      write!(f, "npm:{}/{}", self.nv, sub_path)
-    } else {
-      write!(f, "npm:{}", self.nv)
-    }
-  }
-}
-
-#[derive(Debug, Error, Clone)]
-#[error("Invalid npm package name and version '{text}'. {message}")]
-pub struct NpmPackageNvParseError {
-  pub message: String,
-  pub text: String,
-}
-
-#[derive(
-  Clone, PartialOrd, Ord, PartialEq, Eq, Hash, Serialize, Deserialize,
-)]
-pub struct NpmPackageNv {
-  pub name: String,
-  pub version: Version,
-}
-
-impl std::fmt::Debug for NpmPackageNv {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    // when debugging, it's easier to compare this
-    write!(f, "{}@{}", self.name, self.version)
-  }
-}
-
-impl std::fmt::Display for NpmPackageNv {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    write!(f, "{}@{}", self.name, self.version)
-  }
-}
-
-impl NpmPackageNv {
-  #[allow(clippy::should_implement_trait)]
-  pub fn from_str(nv: &str) -> Result<Self, NpmPackageNvParseError> {
-    monch::with_failure_handling(parse_nv)(nv).map_err(|err| {
-      NpmPackageNvParseError {
-        message: format!("{err:#}"),
-        text: nv.to_string(),
-      }
-    })
-  }
-
-  pub fn scope(&self) -> Option<&str> {
-    if self.name.starts_with('@') {
-      self.name.split('/').next()
-    } else {
-      None
-    }
-  }
-}
-
-fn parse_nv(input: &str) -> monch::ParseResult<NpmPackageNv> {
-  use monch::*;
-
-  fn parse_name(input: &str) -> ParseResult<&str> {
-    if_not_empty(substring(move |input| {
-      for (pos, c) in input.char_indices() {
-        // first character might be a scope, so skip it
-        if pos > 0 && c == '@' {
-          return Ok((&input[pos..], ()));
-        }
-      }
-      ParseError::backtrace()
-    }))(input)
-  }
-
-  fn parse_version(input: &str) -> ParseResult<&str> {
-    if_not_empty(substring(skip_while(|c| !matches!(c, '_' | '/'))))(input)
-  }
-
-  let (input, name) = parse_name(input)?;
-  let (input, _) = ch('@')(input)?;
-  let at_version_input = input;
-  let (input, version) = parse_version(input)?;
-  match Version::parse_from_npm(version) {
-    Ok(version) => Ok((
-      input,
-      NpmPackageNv {
-        name: name.to_string(),
-        version,
-      },
-    )),
-    Err(err) => ParseError::fail(at_version_input, format!("{err:#}")),
-  }
-}
-
 #[derive(Error, Debug, Clone)]
 pub enum NpmPackageReqReferenceParseError {
   #[error("Not an npm specifier.")]
@@ -810,7 +643,7 @@ impl<'de> Deserialize<'de> for NpmPackageReq {
     D: serde::Deserializer<'de>,
   {
     let text = String::deserialize(deserializer)?;
-    match NpmPackageReq::from_str(&text) {
+    match Self::from_str(&text) {
       Ok(req) => Ok(req),
       Err(err) => Err(serde::de::Error::custom(err)),
     }
@@ -1432,41 +1265,6 @@ mod tests {
       // by node semver so we should error
       assert!(parse_npm_version_req(req_text).is_err());
     }
-  }
-
-  #[test]
-  fn npm_package_nv_ref() {
-    let package_nv_ref =
-      NpmPackageNvReference::from_str("npm:package@1.2.3/test").unwrap();
-    assert_eq!(
-      package_nv_ref,
-      NpmPackageNvReference {
-        nv: NpmPackageNv {
-          name: "package".to_string(),
-          version: Version::parse_from_npm("1.2.3").unwrap(),
-        },
-        sub_path: Some("test".to_string())
-      }
-    );
-    assert_eq!(
-      package_nv_ref.as_specifier().as_str(),
-      "npm:/package@1.2.3/test"
-    );
-
-    // no path
-    let package_nv_ref =
-      NpmPackageNvReference::from_str("npm:package@1.2.3").unwrap();
-    assert_eq!(
-      package_nv_ref,
-      NpmPackageNvReference {
-        nv: NpmPackageNv {
-          name: "package".to_string(),
-          version: Version::parse_from_npm("1.2.3").unwrap(),
-        },
-        sub_path: None
-      }
-    );
-    assert_eq!(package_nv_ref.as_specifier().as_str(), "npm:/package@1.2.3");
   }
 
   #[test]
